pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "crewai-instanode-tools"
version = "0.1.0"
description = "CrewAI tools for instanode.dev — provision Postgres + webhooks from CrewAI agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = [
    "crewai",
    "crewai-tools",
    "instanode",
    "postgres",
    "database",
    "webhook",
    "ai-agents",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Database",
]
dependencies = [
    "instanode>=0.2.0",
    "crewai-tools>=0.2",
    "pydantic>=2.0",
]

[project.optional-dependencies]
test = ["pytest>=8.0"]

[project.urls]
Homepage = "https://instanode.dev"
Repository = "https://github.com/InstaNode-dev/crewai-instanode-tools"
"Bug Tracker" = "https://github.com/InstaNode-dev/crewai-instanode-tools/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["crewai_instanode_tools*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
